backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Kjell Boeije <kjellboeije@gmail.com>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
django-rest-framework = "^0.1.0"
channels = "^4.1.0"
django-cors-headers = "^4.3.1"
daphne = "^4.1.2"
channels-redis = "^4.2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
